Support export OTel metrics to be EMF in starndard output
